+++ Sources/Sentry/SentrySpan.m
@@ -74,7 +74,7 @@ SentrySpan ()
     }
 }
 
-- (nullable NSDictionary<NSString *, id> *)data
+- (NSDictionary<NSString *, id> *)data
 {
     @synchronized(_data) {
         return [_data copy];